+++ R/0_imports.R
@@ -35,6 +35,7 @@ utils::globalVariables(
     "denom",                                                           # ratio.R
     "component", "denom", "id", "value",                                 # ica.R
     "training", "x_names", "y_names",                                    # pls.R
-    ".order_1", ".order_2", "role", "skip"                           # recipes.R
+    ".order_1", ".order_2", "role", "skip",                          # recipes.R
+    "dat"                                                        # nnmf_sparse.R
   )
 )